separated programCourse tests from test_app updated wsgi to run programCourse tests
